server/requirements.txt

fastapi>=0.109.0
uvicorn>=0.27.0
python-telegram-bot>=20.7
google-genai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.10.0
jinja2>=3.1.3
requests>=2.31.0
cryptography>=41.0.0
